enhacing...
Map fixes.
player jump implementation changes.
player's overall animation speedup.
save & enhance function fix

+++ Assets/Script/Status.cs
@@ -8,7 +8,7 @@ public class Status : MonoBehaviour
     [SerializeField]
     protected int maxHP;
     [SerializeField]
-    protected int power;
+    protected float power;
     [SerializeField]
     protected float jumpPower;
     [SerializeField]
